Games model and seeding
The freetogame api is a little unreliable. The gamesApiData.json is the last full list I was able to pull from it so I saved that to a file we can populate from. I updated the seed.js to make that work and updated the Games.js model so that we're not tying a game to a specific user. I'm not exactly sure how to pre-populate the db using the free tier on Render.